+++ pkg/resource/resourceConfig.go
@@ -2760,10 +2760,11 @@ func CreatePolicy(rls Rules, policyName, partition string) *Policy {
 	return &plcy
 }
 
-func (cm *AgentCfgMap) Init(n string, ns string, d string, l map[string]string, getEP func(string, string, bool) ([]Member, error)) {
+func (cm *AgentCfgMap) Init(n string, ns string, d string, l map[string]string, annotation string, getEP func(string, string, bool) ([]Member, error)) {
 	cm.Name = n
 	cm.Namespace = ns
 	cm.Data = d
 	cm.Label = l
+	cm.Annotation = annotation
 	cm.GetEndpoints = getEP
 }